Rename `util` to `utils`
Also update examples and docs to import from `utils`.
