+++ bionic/__init__.py
@@ -15,5 +15,6 @@ from .decorators import (  # noqa: F401
 )
 
 from . import protocol  # noqa: F401
+from . import util  # noqa: F401
 
 __version__ = u"0.8.2"
