+++ bionic/cache_api.py
@@ -7,7 +7,7 @@ from functools import total_ordering
 
 from .descriptors.parsing import entity_dnode_from_descriptor
 from .oneline import oneline
-from .util.urls import path_from_url, is_file_url
+from .utils.urls import path_from_url, is_file_url
 
 
 class Cache:
