+++ bionic/core/task_execution.py
@@ -12,7 +12,7 @@ from ..datatypes import ProvenanceDigest, Query, Result
 from ..exception import CodeVersioningError
 from ..oneline import oneline
 from ..persistence import Provenance
-from ..util.misc import single_unique_element
+from ..utils.misc import single_unique_element
 
 
 class TaskRunnerEntry:
