+++ bionic/flow.py
@@ -43,8 +43,8 @@ from .descriptors.parsing import entity_dnode_from_descriptor
 from . import decorators, decoration
 from .filecopier import FileCopier
 from .oneline import oneline
-from .util.gcs import copy_to_gcs
-from .util.misc import (
+from .utils.gcs import copy_to_gcs
+from .utils.misc import (
     group_pairs,
     check_exactly_one_present,
     check_at_most_one_present,
