+++ bionic/provider.py
@@ -29,7 +29,7 @@ from .descriptors.parsing import entity_dnode_from_descriptor
 from .descriptors import ast
 from .exception import EntityComputationError, IncompatibleEntityError
 from .oneline import oneline
-from .util.misc import groups_dict, hash_to_hex
+from .utils.misc import groups_dict, hash_to_hex
 
 import logging
