+++ example/basic_workflow.py
@@ -14,7 +14,7 @@ def x_plus_y(x, y):
 flow = builder.build()
 
 if __name__ == "__main__":
-    bn.util.init_basic_logging()
+    bn.utils.misc.init_basic_logging()
 
     for _, row in flow.get("x_plus_y", "series").reset_index().iterrows():
         print(f"{row['x']} + {row['y']} = {row['x_plus_y']}")
