+++ tests/test_utils.py
@@ -4,7 +4,7 @@ from .helpers import equal_when_sorted
 
 
 def test_group_pairs():
-    from bionic.util.misc import group_pairs
+    from bionic.utils.misc import group_pairs
 
     assert group_pairs([]) == []
     assert group_pairs([1, 2]) == [(1, 2)]
@@ -17,7 +17,7 @@ def test_group_pairs():
 
 
 def test_immutable_sequence():
-    from bionic.util.misc import ImmutableSequence
+    from bionic.utils.misc import ImmutableSequence
 
     class Seq(ImmutableSequence):
         def __init__(self, items):
@@ -48,7 +48,7 @@ def test_immutable_sequence():
 
 
 def test_immutable_mapping():
-    from bionic.util.misc import ImmutableMapping
+    from bionic.utils.misc import ImmutableMapping
 
     class Mapping(ImmutableMapping):
         def __init__(self, values_by_key):
@@ -84,7 +84,7 @@ def test_immutable_mapping():
     assert Mapping({"b": 1, "a": 2}) >= mapping
 
 
-# This function is not in util.misc.py but it's convenient to test it here too.
+# This function is not in utils.misc.py but it's convenient to test it here too.
 def test_oneline():
     from bionic.oneline import oneline
 
@@ -113,7 +113,7 @@ def test_oneline():
 
 
 def test_clean_docstring():
-    from bionic.util.misc import rewrap_docstring
+    from bionic.utils.misc import rewrap_docstring
 
     assert rewrap_docstring("") == ""
     assert rewrap_docstring("test") == "test"
@@ -233,7 +233,7 @@ def test_clean_docstring():
     assert rewrap_docstring(doc) == "test\n- one\ntwo"
 
 
-# These functions are not in util.misc.py but it's convenient to test them here too.
+# These functions are not in utils.misc.py but it's convenient to test them here too.
 def test_longest_regex_prefix():
     from .helpers import longest_regex_prefix_match
